Test reencoding of GIFs
