+++ Tests/GIFTests/XCTestManifests.swift
@@ -5,7 +5,7 @@ public func allTests() -> [XCTestCaseEntry] {
     return [
         testCase(LzwCoderTests.allTests),
         testCase(BitDataTests.allTests),
-        testCase(GIFDecoderTests.allTests),
+        testCase(GIFCoderTests.allTests),
         testCase(PackedFieldByteTests.allTests)
     ]
 }
